Move package install/update under xpkg
